Abstract Project symbol table logic in a dedicated class (#729)
