+++ python-frontend/src/main/java/org/sonar/plugins/python/api/PythonVisitorContext.java
@@ -23,13 +23,11 @@ import com.sonar.sslr.api.RecognitionException;
 import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
-import java.util.Set;
 import javax.annotation.CheckForNull;
 import javax.annotation.Nullable;
 import org.sonar.plugins.python.api.PythonCheck.PreciseIssue;
-import org.sonar.plugins.python.api.symbols.Symbol;
 import org.sonar.plugins.python.api.tree.FileInput;
+import org.sonar.python.semantic.ProjectLevelSymbolTable;
 import org.sonar.python.semantic.SymbolTableBuilder;
 
 public class PythonVisitorContext {
@@ -50,12 +48,12 @@ public class PythonVisitorContext {
     symbolTableBuilder.visitFileInput(rootTree);
   }
 
-  public PythonVisitorContext(FileInput rootTree, PythonFile pythonFile, @Nullable File workingDirectory, String packageName, Map<String, Set<Symbol>> globalSymbols) {
+  public PythonVisitorContext(FileInput rootTree, PythonFile pythonFile, @Nullable File workingDirectory, String packageName, ProjectLevelSymbolTable projectLevelSymbolTable) {
     this.rootTree = rootTree;
     this.pythonFile = pythonFile;
     this.workingDirectory = workingDirectory;
     this.parsingException = null;
-    new SymbolTableBuilder(packageName, pythonFile, globalSymbols).visitFileInput(rootTree);
+    new SymbolTableBuilder(packageName, pythonFile, projectLevelSymbolTable).visitFileInput(rootTree);
   }
 
   public PythonVisitorContext(PythonFile pythonFile, RecognitionException parsingException) {
