+++ python-frontend/src/main/java/org/sonar/python/TestPythonVisitorRunner.java
@@ -25,17 +25,14 @@ import java.io.IOException;
 import java.net.URI;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
-import java.util.Collections;
 import java.util.List;
-import java.util.Map;
-import java.util.Set;
 import javax.annotation.Nullable;
 import org.sonar.plugins.python.api.PythonCheck;
 import org.sonar.plugins.python.api.PythonFile;
 import org.sonar.plugins.python.api.PythonVisitorContext;
-import org.sonar.plugins.python.api.symbols.Symbol;
 import org.sonar.plugins.python.api.tree.FileInput;
 import org.sonar.python.parser.PythonParser;
+import org.sonar.python.semantic.ProjectLevelSymbolTable;
 import org.sonar.python.semantic.SymbolUtils;
 import org.sonar.python.tree.PythonTreeMaker;
 
@@ -59,28 +56,27 @@ public class TestPythonVisitorRunner {
   }
 
   public static PythonVisitorContext createContext(File file, @Nullable File workingDirectory) {
-    return createContext(file, workingDirectory, "", Collections.emptyMap());
+    return createContext(file, workingDirectory, "", ProjectLevelSymbolTable.empty());
   }
 
-  public static PythonVisitorContext createContext(File file, @Nullable File workingDirectory, String packageName, Map<String, Set<Symbol>> globalSymbols) {
+  public static PythonVisitorContext createContext(File file, @Nullable File workingDirectory, String packageName, ProjectLevelSymbolTable projectLevelSymbolTable) {
     PythonParser parser = PythonParser.create();
     TestPythonFile pythonFile = new TestPythonFile(file);
     AstNode astNode = parser.parse(pythonFile.content());
     FileInput rootTree = new PythonTreeMaker().fileInput(astNode);
-    return new PythonVisitorContext(rootTree, pythonFile, workingDirectory, packageName, globalSymbols);
+    return new PythonVisitorContext(rootTree, pythonFile, workingDirectory, packageName, projectLevelSymbolTable);
   }
 
-  public static Map<String, Set<Symbol>> globalSymbols(List<File> files, File baseDir) {
-    Map<String, Set<Symbol>> globalSymbols = SymbolUtils.externalModulesSymbols();
+  public static ProjectLevelSymbolTable globalSymbols(List<File> files, File baseDir) {
+    ProjectLevelSymbolTable projectLevelSymbolTable = ProjectLevelSymbolTable.from(SymbolUtils.externalModulesSymbols());
     for (File file : files) {
       TestPythonFile pythonFile = new TestPythonFile(file);
       AstNode astNode = PythonParser.create().parse(pythonFile.content());
       FileInput astRoot = new PythonTreeMaker().fileInput(astNode);
       String packageName = pythonPackageName(file, baseDir);
-      String fullyQualifiedModuleName = SymbolUtils.fullyQualifiedModuleName(packageName, pythonFile.fileName());
-      globalSymbols.put(fullyQualifiedModuleName, SymbolUtils.globalSymbols(astRoot, packageName, pythonFile));
+      projectLevelSymbolTable.addModule(astRoot, packageName, pythonFile);
     }
-    return globalSymbols;
+    return projectLevelSymbolTable;
   }
 
   private static class TestPythonFile implements PythonFile {
